The Whole Game - V3 [Lights + Assets]
+ Added most of the finished assets.
+ Added particle systems (all don't really work well at the moment).

Bugs/Problems:
The particle systems need to be coded in, at the moment one of the scripts is generating a null reference error.

The lights system is a bit finnicky at the moment, the torch particles aren't tied properly to the original lights which is making things a bit confusing.